Figaro Working, but CSRF being thrown.
I Finally got figaro working. I don't know exactly what did it, but messing around with omniauth.rb eventually did it... I think. I also created the support necessary on the Google project to support hitting their site. The only problem I'm facing now is that a CSRF error is popping up when we try to allow the app to do it's thing. I read online that the issue is with the domains matching between google and the app. I have fixed that; didn't fix the issue.