Added the Seed file.
I started with a dream driven seed file. Then I created the necessary models when the seed was telling me to do so. Then I formed the relationships and created 'make_new' methods for each of the objects. It all came together with the Pokegem wrapper for this API. Very easily I can grab a Pokeobject and manipulate it as an openstruct.