Added twitter image and updated user.
The twitter image was very easy to implement, but I don't see how it would be useful. I also had to update the user model because twitter does not provide as much as the google OAuth. I plan on spending a bit making a pixellated twitter bird on the front page, but that is low priority.