TDD'd base team functionality.
Current User gave me some issues when trying to mock the twitter omniauth, but after I got it working it reaffirmed how powerful it is. Conditionals in views are a thing of the past. I wrote and passed a feature test which lead into a few model tests and a couple integration tests. Feels good to be TDDing finally. Also I don't like css.