Added mouse hover/swap stats for index.
The angular wasn't too hard, but the css was a bit of an issue. I finally have it working by just reusing a lot of the css for the original display. I don't know how that will hold up once I style the stats and comparisons which are up next.