Tweaked stat comparison display through decorators.
CSS is not the greatest, but I've finally got a prettier looking site. The stat display as colored text is a huge improvement from the arrow system I was clumsily trying to implement. I have my gf to thank for that. The next big step is trying to decide how to style the team page better. For now I'm going to work on the easy polish stuff: font, mouse cursor, background image denoisifying, whole image being a button.